Ref: extracting lastResortHandler into a dedicated file.
